feat: delete 확인창 및 응답 추가

+++ admin/src/components/Table/ManagerTableColumns.jsx
@@ -27,16 +27,20 @@ export const ManagerTableColumns = (info) => [
 //     console.error("비밀번호 변경 에러:", error);
 //   }
 // };
-
+const redirectU = "https://myadmin.wheelgo.net/manager";
 const handleDeleteAccount = async (email) => {
-  try {
-    console.log("삭제 클릭: ", email);
+  const confirmDelete = window.confirm("삭제하시겠습니까?");
+  if (confirmDelete) {
+    try {
+      console.log("삭제 클릭: ", email);
 
-    const apiUrl = `https://devapi.wheelgo.net/api/v1/auth/withdraw?id=${email}`;
-    const response = await axios.delete(apiUrl);
+      const apiUrl = `https://devapi.wheelgo.net/api/v1/auth/withdraw?id=${email}`;
+      const response = await axios.delete(apiUrl);
 
-    console.log("삭제 응답 데이터:", response.data);
-  } catch (error) {
-    console.error("계정 삭제 에러:", error);
+      console.log("삭제 응답 데이터:", response.data);
+      window.location.href = redirectU;
+    } catch (error) {
+      console.error("계정 삭제 에러:", error);
+    }
   }
 };
